fix: preserve semantic admission evidence
