+++ backend-idea/src/main/kotlin/io/github/amichne/kast/idea/ReferenceSearchClock.kt
@@ -1,0 +1,9 @@
+package io.github.amichne.kast.idea
+
+internal fun interface ReferenceSearchClock {
+    fun nanoTime(): Long
+
+    companion object {
+        val System: ReferenceSearchClock = ReferenceSearchClock { java.lang.System.nanoTime() }
+    }
+}
